Clarify multimodal ownership boundaries
