+++ vllm_metal/multimodal/feature_spec.py
@@ -1,0 +1,8 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Feature-spec aliases for multimodal inputs."""
+
+from __future__ import annotations
+
+from vllm.multimodal.inputs import MultiModalFeatureSpec, PlaceholderRange
+
+__all__ = ["MultiModalFeatureSpec", "PlaceholderRange"]
