+++ vllm_metal/multimodal/qwen3_vl/__init__.py
@@ -1,0 +1,8 @@
+# SPDX-License-Identifier: Apache-2.0
+"""Qwen3-VL multimodal adapter helpers."""
+
+from __future__ import annotations
+
+from vllm_metal.multimodal.qwen3_vl.adapter import Qwen3VLMultimodalAdapter
+
+__all__ = ["Qwen3VLMultimodalAdapter"]
