+++ vllm_metal/v1/mm/__init__.py
@@ -1,0 +1,8 @@
+# SPDX-License-Identifier: Apache-2.0
+"""v1 multimodal runtime helpers for vLLM Metal."""
+
+from __future__ import annotations
+
+from vllm_metal.v1.mm.encoder_cache import EncoderCache
+
+__all__ = ["EncoderCache"]
